TRAFFICSIGN/requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=8.3.0
